requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Database
motor==3.3.2
pymongo==4.6.0

# Document processing
pypdf==3.17.1
python-docx==1.1.0

# AI/ML
google-generativeai==0.3.2

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1